+++ src/test/java/com/mihai/writer/CellStyleCreatorXlsxTest.java
@@ -84,7 +84,7 @@ class CellStyleCreatorXlsxTest {
     public void fontApplied() {
         StyleFont font = StyleFont.builder()
                 .name("Calibri")
-                .size((short) 18)
+                .size(18)
                 .color(new StyleColor(1, 2, 3))
                 .bold(true)
                 .italic(true)
